feat(frontend): the account header offers both verbs, and stops living in a 2,700-line file
TWO VERBS, NOT ONE. "Log activity" was the only action on a company, and
setting what happens NEXT was two clicks inside it behind a type picker —
which is how accounts collect notes and no follow-ups. "Add task" is its own
button now, on the same form. Neither is offered on an archived company: the
server refuses the write, so the button would only open a modal that fails on
save.

"Write email" is deliberately absent. ComposeModal takes an activityId, so an
account-started email has no contract to call yet; a button that opens a
composer with nothing to send from is a promise this page cannot keep. It joins
the group when the account-started draft/send contract does.

THE HEADER MOVES OUT. organizations.tsx carried the list screen, the enrichment
tools, the evidence cards, the chronology and the record header at once, and
the V2 work adds to every one of them. The header — the action group, the
edit/merge/archive overflow, the inline lifecycle and owner controls, and the
pulse line — is now companyheader.tsx. 2,733 lines becomes 2,237 + 537.

The overlay-refusal coverage gate follows the merge call to its new file rather
than being relaxed: it exists to prove every mirrored write translates its
refusal, and a gate that stops looking is a gate that stops holding.
